# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LinkedIn Intelligence & Job Scraper API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type RunScraperRequestScrapeType =
  (typeof RunScraperRequestScrapeType)[keyof typeof RunScraperRequestScrapeType];

export const RunScraperRequestScrapeType = {
  posts: "posts",
  jobs: "jobs",
  both: "both",
} as const;

export interface RunScraperRequest {
  /** Keywords to search for. Uses saved keywords if empty. */
  keywords?: string[];
  /** Max results per keyword */
  maxResults?: number;
  scrapeType?: RunScraperRequestScrapeType;
}

export interface RunScraperResponse {
  runId: string;
  status: string;
  message: string;
  keywords: string[];
}

export type ScraperRunStatusStatus =
  (typeof ScraperRunStatusStatus)[keyof typeof ScraperRunStatusStatus];

export const ScraperRunStatusStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface ScraperRunStatus {
  runId: string;
  status: ScraperRunStatusStatus;
  startedAt: string;
  completedAt?: string | null;
  totalFetched: number;
  totalProcessed: number;
  totalDuplicates: number;
  totalErrors: number;
  keywords: string[];
  errorMessage?: string | null;
}

export interface ScraperRunList {
  runs: ScraperRunStatus[];
  total: number;
}

export interface Keyword {
  id: number;
  keyword: string;
  createdAt: string;
  active: boolean;
}

export interface KeywordList {
  keywords: Keyword[];
}

export interface CreateKeywordRequest {
  keyword: string;
}

export interface JobRecord {
  id: number;
  postId: string;
  role?: string | null;
  companyName?: string | null;
  location?: string | null;
  primarySkills?: string | null;
  secondarySkills?: string | null;
  mustToHave?: string | null;
  yearsOfExperience?: string | null;
  lookingForCollegeStudents?: string | null;
  intern?: string | null;
  salaryPackage?: string | null;
  email?: string | null;
  phone?: string | null;
  hiringIntent?: string | null;
  authorName?: string | null;
  authorLinkedinUrl?: string | null;
  postUrl?: string | null;
  datePosted?: string | null;
  dateProcessed?: string | null;
  keywordMatched?: string | null;
  rawText?: string | null;
  sheetRowId?: number | null;
}

export interface JobList {
  jobs: JobRecord[];
  total: number;
}

export type JobStatsTopCompaniesItem = {
  company: string;
  count: number;
};

export type JobStatsTopKeywordsItem = {
  keyword: string;
  count: number;
};

export interface JobStats {
  totalJobs: number;
  totalRuns: number;
  jobsLast24h: number;
  jobsLast7d: number;
  topCompanies: JobStatsTopCompaniesItem[];
  topKeywords: JobStatsTopKeywordsItem[];
}

export type ListScraperRunsParams = {
  limit?: number;
  offset?: number;
};

export type ListJobsParams = {
  limit?: number;
  offset?: number;
  keyword?: string;
  company?: string;
};