# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LinkedIn Intelligence & Job Scraper API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Trigger a scraping run
 */
export const runScraperBodyMaxResultsDefault = 100;
export const runScraperBodyScrapeTypeDefault = `both`;

export const RunScraperBody = zod.object({
  keywords: zod
    .array(zod.string())
    .optional()
    .describe("Keywords to search for. Uses saved keywords if empty."),
  maxResults: zod
    .number()
    .default(runScraperBodyMaxResultsDefault)
    .describe("Max results per keyword"),
  scrapeType: zod
    .enum(["posts", "jobs", "both"])
    .default(runScraperBodyScrapeTypeDefault),
});

export const RunScraperResponse = zod.object({
  runId: zod.string(),
  status: zod.string(),
  message: zod.string(),
  keywords: zod.array(zod.string()),
});

/**
 * @summary Get scraper run status
 */
export const GetScraperStatusParams = zod.object({
  runId: zod.coerce.string(),
});

export const GetScraperStatusResponse = zod.object({
  runId: zod.string(),
  status: zod.enum(["pending", "running", "completed", "failed"]),
  startedAt: zod.coerce.date(),
  completedAt: zod.coerce.date().nullish(),
  totalFetched: zod.number(),
  totalProcessed: zod.number(),
  totalDuplicates: zod.number(),
  totalErrors: zod.number(),
  keywords: zod.array(zod.string()),
  errorMessage: zod.string().nullish(),
});

/**
 * @summary List recent scraping runs
 */
export const listScraperRunsQueryLimitDefault = 20;
export const listScraperRunsQueryOffsetDefault = 0;

export const ListScraperRunsQueryParams = zod.object({
  limit: zod.coerce.number().default(listScraperRunsQueryLimitDefault),
  offset: zod.coerce.number().default(listScraperRunsQueryOffsetDefault),
});

export const ListScraperRunsResponse = zod.object({
  runs: zod.array(
    zod.object({
      runId: zod.string(),
      status: zod.enum(["pending", "running", "completed", "failed"]),
      startedAt: zod.coerce.date(),
      completedAt: zod.coerce.date().nullish(),
      totalFetched: zod.number(),
      totalProcessed: zod.number(),
      totalDuplicates: zod.number(),
      totalErrors: zod.number(),
      keywords: zod.array(zod.string()),
      errorMessage: zod.string().nullish(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary List all keywords
 */
export const ListKeywordsResponse = zod.object({
  keywords: zod.array(
    zod.object({
      id: zod.number(),
      keyword: zod.string(),
      createdAt: zod.coerce.date(),
      active: zod.boolean(),
    }),
  ),
});

/**
 * @summary Add a keyword
 */
export const CreateKeywordBody = zod.object({
  keyword: zod.string(),
});

/**
 * @summary Delete a keyword
 */
export const DeleteKeywordParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteKeywordResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary List processed job records
 */
export const listJobsQueryLimitDefault = 50;
export const listJobsQueryOffsetDefault = 0;

export const ListJobsQueryParams = zod.object({
  limit: zod.coerce.number().default(listJobsQueryLimitDefault),
  offset: zod.coerce.number().default(listJobsQueryOffsetDefault),
  keyword: zod.coerce.string().optional(),
  company: zod.coerce.string().optional(),
});

export const ListJobsResponse = zod.object({
  jobs: zod.array(
    zod.object({
      id: zod.number(),
      postId: zod.string(),
      role: zod.string().nullish(),
      companyName: zod.string().nullish(),
      location: zod.string().nullish(),
      primarySkills: zod.string().nullish(),
      secondarySkills: zod.string().nullish(),
      mustToHave: zod.string().nullish(),
      yearsOfExperience: zod.string().nullish(),
      lookingForCollegeStudents: zod.string().nullish(),
      intern: zod.string().nullish(),
      salaryPackage: zod.string().nullish(),
      email: zod.string().nullish(),
      phone: zod.string().nullish(),
      hiringIntent: zod.string().nullish(),
      authorName: zod.string().nullish(),
      authorLinkedinUrl: zod.string().nullish(),
      postUrl: zod.string().nullish(),
      datePosted: zod.string().nullish(),
      dateProcessed: zod.string().nullish(),
      keywordMatched: zod.string().nullish(),
      rawText: zod.string().nullish(),
      sheetRowId: zod.number().nullish(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Get job statistics
 */
export const GetJobStatsResponse = zod.object({
  totalJobs: zod.number(),
  totalRuns: zod.number(),
  jobsLast24h: zod.number(),
  jobsLast7d: zod.number(),
  topCompanies: zod.array(
    zod.object({
      company: zod.string(),
      count: zod.number(),
    }),
  ),
  topKeywords: zod.array(
    zod.object({
      keyword: zod.string(),
      count: zod.number(),
    }),
  ),
});