# lib/api-zod/src/generated/types/jobList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LinkedIn Intelligence & Job Scraper API
 * OpenAPI spec version: 0.1.0
 */
import type { JobRecord } from "./jobRecord";

export interface JobList {
  jobs: JobRecord[];
  total: number;
}