# lib/api-zod/src/generated/types/jobRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LinkedIn Intelligence & Job Scraper API
 * OpenAPI spec version: 0.1.0
 */

export interface JobRecord {
  id: number;
  postId: string;
  role?: string | null;
  companyName?: string | null;
  location?: string | null;
  primarySkills?: string | null;
  secondarySkills?: string | null;
  mustToHave?: string | null;
  yearsOfExperience?: string | null;
  lookingForCollegeStudents?: string | null;
  intern?: string | null;
  salaryPackage?: string | null;
  email?: string | null;
  phone?: string | null;
  hiringIntent?: string | null;
  authorName?: string | null;
  authorLinkedinUrl?: string | null;
  postUrl?: string | null;
  datePosted?: string | null;
  dateProcessed?: string | null;
  keywordMatched?: string | null;
  rawText?: string | null;
  sheetRowId?: number | null;
}