# lib/api-zod/src/generated/types/runScraperRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LinkedIn Intelligence & Job Scraper API
 * OpenAPI spec version: 0.1.0
 */
import type { RunScraperRequestScrapeType } from "./runScraperRequestScrapeType";

export interface RunScraperRequest {
  /** Keywords to search for. Uses saved keywords if empty. */
  keywords?: string[];
  /** Max results per keyword */
  maxResults?: number;
  scrapeType?: RunScraperRequestScrapeType;
}