# lib/api-zod/src/generated/types/scraperRunList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LinkedIn Intelligence & Job Scraper API
 * OpenAPI spec version: 0.1.0
 */
import type { ScraperRunStatus } from "./scraperRunStatus";

export interface ScraperRunList {
  runs: ScraperRunStatus[];
  total: number;
}